# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Siuji TOEFL exam application.",
        "title": "Siuji API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/questions/{question_public_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a question with its options and correct answer.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Get question detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Question public ID",
                        "name": "question_public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.QuestionDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update question text/passage. audio/image are optional — omit to keep the existing files.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Update question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Question public ID",
                        "name": "question_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Question text",
                        "name": "question",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Reading passage",
                        "name": "passage",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "New audio file (optional)",
                        "name": "audio",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "New image file (optional)",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.QuestionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Delete question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Question public ID",
                        "name": "question_public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
...
        "/sections/{section_public_id}/questions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new question in a section. audio and image files are optional depending on section type.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Create question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Section public ID",
                        "name": "section_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Question text",
                        "name": "question",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Reading passage (for Reading section)",
                        "name": "passage",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Audio file (for Listening section)",
                        "name": "audio",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Image file (optional)",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.QuestionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
        "/sections/{section_public_id}/questions/reorder": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Reorder questions within a section",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Section public ID",
                        "name": "section_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ordered list of question public IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReorderQuestionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.OptionResponse": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "option_text": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "public_id": {
                    "type": "string"
                }
            }
        },
...
        "model.QuestionDetailResponse": {
            "type": "object",
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "correct_option_public_id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OptionResponse"
                    }
                },
                "passage": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "model.QuestionResponse": {
            "type": "object",
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "passage": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
...
        "model.ReorderQuestionsRequest": {
            "type": "object",
            "required": [
                "question_public_ids"
            ],
            "properties": {
                "question_public_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.ResponseNoData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "TempToken": {
            "description": "Type \"Bearer\" followed by a space and temp JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}